[refactor] credit API 스키마 경로를 피처 모듈로 이동 및 업데이트
credit 관련 API 스키마 타입 정의를 공통 타입 경로에서 credit 피처 모듈로 이동하고,
모든 API 라우트 및 컴포넌트에서 새로운 경로를 참조하도록 업데이트함.
DB 스키마 경로도 함께 수정함.
