+++ src/features/flow/components/sharing/business/workflow-share-detail-content.tsx
@@ -8,13 +8,13 @@ import { toast } from "sonner";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
+import type { CreditSummaryResponse } from "@/features/credit/types/credit-apis";
 import { WorkflowSharePurchaseDialog } from "@/features/flow/components/sharing/business/workflow-share-purchase-dialog";
 import { WorkflowGraphPreview } from "@/features/flow/components/sharing/ui/workflow-graph-preview";
 import { WorkflowShareHero } from "@/features/flow/components/sharing/ui/workflow-share-hero";
 import { WorkflowShareSection } from "@/features/flow/components/sharing/ui/workflow-share-section";
 import type { WorkflowShareDetail } from "@/features/flow/types/workflow-sharing";
 import { api } from "@/lib/api-client";
-import type { CreditSummaryResponse } from "@/types/credit/credit-schemas";
 
 interface WorkflowShareDetailContentProps {
   share: WorkflowShareDetail;
